requirements.txt
aiohttp==3.5.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
autopep8==1.4.4
beautifulsoup4==4.8.1
bottle==0.12.17
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
discord.py==1.0.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
lxml==4.4.2
mccabe==0.6.1
multidict==4.5.2
pycodestyle==2.5.0
pylint==2.3.1
requests==2.22.0
rope==0.14.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.5
spotipy==2.4.4
typed-ast==1.4.0
urllib3==1.25.3
websockets==6.0
wrapt==1.11.2
yarl==1.3.0
